refactor(parser): 重构FBX解析器模块结构
将FBX解析器拆分为多个子模块，包括连接解析、变形器解析、模型构建等
更新文档状态并移除旧文件引用，所有功能已迁移至新结构
修复类型定义并优化代码可维护性
