+++ docs/fbx-tree-parser-execution-steps.md
@@ -163,17 +163,22 @@ pnpm lint      # 代码风格检查
 
 ## 当前状态
 
-- [ ] 步骤1：创建目录结构
-- [ ] 步骤2：类型守护工具整理
-- [ ] 步骤3：连接解析功能
-- [ ] 步骤4：图像解析功能
-- [ ] 步骤5：纹理解析功能
-- [ ] 步骤6：材质解析功能
-- [ ] 步骤7：变形器解析功能
-- [ ] 步骤8：场景构建功能
-- [ ] 步骤9：模型解析功能
-- [ ] 步骤10：变换工具
-- [ ] 步骤11：统一导出
-- [ ] 步骤12：新FBXTreeParser
-- [ ] 步骤13：更新主加载器
-- [ ] 步骤14：测试验证
+- [x] 步骤1：创建目录结构
+- [x] 步骤2：类型守护工具整理
+- [x] 步骤3：连接解析功能
+- [x] 步骤4：图像解析功能
+- [x] 步骤5：纹理解析功能
+- [x] 步骤6：材质解析功能
+- [x] 步骤7：变形器解析功能
+- [x] 步骤8：场景构建功能
+- [x] 步骤9：模型解析功能
+- [x] 步骤10：变换工具
+- [x] 步骤11：统一导出
+- [x] 步骤12：新FBXTreeParser
+- [x] 步骤13：更新主加载器
+- [x] 步骤14：测试验证（已通过 TS/Lint；Demo 正常）
+
+## 变更说明（附加）
+
+- 已完全删除旧文件 `src/parse/FBX-tree-parser.ts`，实现迁移至 `src/parse/tree/FBXTreeParser.ts`。
+- 所有引用已统一到 `parse/tree` 入口或 `parse/tree/index.ts` 导出。
