--- src/parse/FBX-tree-parser.ts
@@ -1,1209 +1,0 @@
-import type {
-  BufferGeometry,
-  EulerOrder,
-  LoadingManager,
-  TextureLoader,
-  MeshStandardMaterial } from 'three';
-import {
-  AmbientLight,
-  Bone,
-  Color,
-  ColorManagement,
-  DirectionalLight,
-  Group,
-  Line,
-  LineBasicMaterial,
-  Loader,
-  MathUtils,
-  Matrix4,
-  Mesh,
-  Object3D,
-  PerspectiveCamera,
-  PointLight,
-  PropertyBinding,
-  Skeleton,
-  SkinnedMesh,
-  SpotLight,
-  SRGBColorSpace,
-  Vector3,
-} from 'three';
-import { BaseParser, type IParsingContext, type FBXConnectionNode, type FBXConnectionReference, type FBXLightNodeAttribute, type FBXMeshNode, type FBXModelNode, type FBXMorphTarget, type FBXRawTargets, type FBXSkeleton, type RawBone, type UserDataTransform, type FBXSceneParserOptions } from '../types';
-import type {
-  IFBXTreeParser,
-  FBXTreeParserInput,
-  FBXTreeParserResult,
-  SceneInfo,
-  GeometryGroup,
-  GeometryWithDeformer,
-  SceneParseResult,
-  ModelInfo,
-  Object3DWithID,
-  BoneWithID,
-} from '../types';
-import type { GeometryInfo } from '../types/parsers/type-guards';
-import {
-  extractNodeId,
-  extractMatrixArray,
-  extractNumberArray,
-  extractMaterialValue,
-  hasTarget,
-} from '../types/parsers/type-guards';
-import { generateTransform, getEulerOrder } from './utils';
-import { AnimationParser } from './FBX-animation-parser';
-import { GeometryParser } from './geometry';
-import { parseConnections } from './tree/connections';
-import { parseImages } from './tree/image-parser';
-import { parseTextures } from './tree/texture-parser';
-import { parseMaterials } from './tree/material-parser';
-
-// Parse the FBXTree object returned by the BinaryParser or TextParser and return a Group
-export class FBXTreeParser extends BaseParser<FBXTreeParserInput, Promise<FBXTreeParserResult>> implements IFBXTreeParser {
-  textureLoader: TextureLoader;
-  manager: LoadingManager;
-  defaultMaterialIndex = -1;
-  private sceneInfo: SceneInfo = {
-    isPBR: true,
-    isTextured: true,
-    isUVMapped: true,
-    polygons: 0,
-    quads: 0,
-    topology: 'Quad',
-    triangles: 0,
-    vertices: 0,
-  };
-
-  constructor (textureLoader: TextureLoader, manager: LoadingManager, context: IParsingContext) {
-    super(context);
-    this.textureLoader = textureLoader;
-    this.manager = manager;
-  }
-
-  async parse (_input: FBXTreeParserInput, _context: IParsingContext): Promise<FBXTreeParserResult> {
-    this.context.connections = parseConnections(this.context.fbxTree);
-
-    const images = parseImages(this.context.fbxTree);
-
-    const textures = await parseTextures(
-      this.context.fbxTree,
-      this.context.connections,
-      images,
-      this.manager,
-      this.textureLoader
-    );
-    const materialMap = parseMaterials(
-      this.context.fbxTree,
-      this.context.connections,
-      textures,
-      this.defaultMaterialIndex
-    );
-    const deformers = this.parseDeformers();
-    const { geoInfoMap, geometryMap } = new GeometryParser(this.context).parse(deformers, this.context);
-
-    return this.parseScene({
-      deformers,
-      geoInfoMap,
-      geometryMap,
-      materialMap,
-    });
-  }
-
-  // Parse nodes in context.fbxTree.Objects.Deformer
-  // Deformer node can contain skinning or Vertex Cache animation data, however only skinning is supported here
-  // Generates map of Skeleton-like objects for use later when generating and binding skeletons.
-  parseDeformers () {
-    const skeletons: Record<string, FBXSkeleton> = {};
-    const morphTargets: Record<string, FBXMorphTarget> = {};
-    const objects = this.context.fbxTree.Objects;
-
-    if (!objects) {
-      throw new Error('No objects found in fbxTree.');
-    }
-    if ('Deformer' in objects) {
-      const DeformerNodes = objects.Deformer;
-
-      for (const nodeID in DeformerNodes) {
-        const deformerNode = DeformerNodes[nodeID];
-
-        const relationships = this.context.connections.get(parseInt(nodeID));
-
-        if (!relationships) {
-          throw new Error('No relationships found for nodeID: ' + nodeID);
-        }
-
-        if (deformerNode?.attrType === 'Skin') {
-          const skeleton = this.parseSkeleton(
-            relationships,
-            DeformerNodes as Record<number, FBXMeshNode>,
-          );
-
-          skeleton.ID = nodeID;
-
-          if (relationships.parents.length > 1) {
-            console.warn(
-              'THREE.FBXLoader: skeleton attached to more than one geometry is not supported.',
-            );
-          }
-          skeleton.geometryID = (relationships.parents[0]).ID;
-
-          skeletons[nodeID] = skeleton;
-        } else if (deformerNode?.attrType === 'BlendShape') {
-          const morphTarget: FBXMorphTarget = {
-            id: nodeID,
-            rawTargets: undefined,
-            skeleton: undefined,
-          };
-
-          morphTarget.rawTargets = this.parseMorphTargets(
-            relationships,
-            DeformerNodes as Record<number, FBXMeshNode>,
-          );
-          morphTarget.id = nodeID;
-
-          if (relationships.parents.length > 1) {
-            console.warn(
-              'THREE.FBXLoader: morph target attached to more than one geometry is not supported.',
-            );
-          }
-
-          morphTargets[nodeID] = morphTarget;
-        }
-      }
-    }
-
-    return {
-      morphTargets: morphTargets,
-      skeletons: skeletons,
-    };
-  }
-
-  // Parse single nodes in context.fbxTree.Objects.Deformer
-  // The top level skeleton node has type 'Skin' and sub nodes have type 'Cluster'
-  // Each skin node represents a skeleton and each cluster node represents a bone
-  parseSkeleton (
-    relationships: FBXConnectionNode,
-    deformerNodes: Record<number, FBXMeshNode>,
-  ): FBXSkeleton {
-    const rawBones: RawBone[] = [];
-
-    relationships.children.forEach(child => {
-      const boneNode = deformerNodes[child.ID];
-
-      if (!boneNode || !boneNode.TransformLink) {
-        throw new Error('THREE.FBXLoader: No bone node found for child ID: ' + child.ID);
-      }
-
-      if (boneNode.attrType !== 'Cluster') {
-        return;
-      }
-
-      const rawBone: RawBone = {
-        ID: child.ID,
-        indices: [],
-        transformLink: new Matrix4().fromArray(boneNode.TransformLink.a),
-        weights: [],
-        // transform: new Matrix4().fromArray( boneNode.Transform.a ),
-        // linkMode: boneNode.Mode,
-      };
-
-      if (boneNode.Indexes && boneNode.Weights) {
-        rawBone.indices = boneNode.Indexes.a;
-        rawBone.weights = boneNode.Weights.a;
-      }
-
-      rawBones.push(rawBone);
-    });
-
-    return {
-      bones: [],
-      geometryID: 0,
-      ID: '',
-      rawBones: rawBones,
-    };
-  }
-
-  // The top level morph deformer node has type "BlendShape" and sub nodes have type "BlendShapeChannel"
-  parseMorphTargets (
-    relationships: FBXConnectionNode,
-    deformerNodes: Record<number, FBXMeshNode>,
-  ): FBXRawTargets[] | undefined {
-    const rawMorphTargets: FBXRawTargets[] = [];
-    const connections = this.context.connections;
-
-    if (!connections) {
-      throw new Error('Global connections is undefined');
-    }
-
-    for (let i = 0; i < relationships.children.length; i++) {
-      const child = relationships.children[i];
-
-      const morphTargetNode = deformerNodes[child.ID];
-      const nodeName = morphTargetNode.attrName || 'target' + i;
-      const nodeID = morphTargetNode.id || i;
-      const rawMorphTarget: FBXRawTargets = {
-        fullWeights: morphTargetNode.FullWeights.a,
-        geoID: 0,
-        id: nodeID,
-        initialWeight: morphTargetNode.DeformPercent,
-        name: nodeName,
-      };
-
-      if (morphTargetNode.attrType !== 'BlendShapeChannel') {
-        return;
-      }
-      let id = child.ID;
-
-      if (typeof child.ID !== 'number') {
-        id = parseInt(child.ID);
-      }
-      rawMorphTarget.geoID
-        = connections.get(id)?.children.filter((child: FBXConnectionReference) => {
-          return child.relationship === undefined;
-        })[0]?.ID || 0;
-
-      rawMorphTargets.push(rawMorphTarget);
-    }
-
-    return rawMorphTargets;
-  }
-
-  // create the main Group() to be returned by the loader
-  parseScene (option: FBXSceneParserOptions): SceneParseResult {
-    const {
-      deformers,
-      geoInfoMap = new Map(),
-      geometryMap = new Map(),
-      materialMap = new Map(),
-    } = option;
-
-    this.context.sceneGraph = new Group();
-    const fbxTree = this.context.fbxTree;
-    const connections = this.context.connections;
-
-    if (!fbxTree || !fbxTree.Objects || !connections) {
-      throw new Error('Global FBXTree or fbxTree.Objects or global connections is undefined');
-    }
-
-    const { modelInfoMap, modelMap } = this.parseWireFrameModels(
-      deformers.skeletons,
-      geometryMap,
-      materialMap,
-      geoInfoMap,
-    );
-
-    const modelNodes = fbxTree.Objects.Model;
-
-    if (!modelNodes) {
-      throw new Error('Model nodes are undefined');
-    }
-    modelMap.forEach(model => {
-      const modelWithID = model as Object3DWithID;
-      const modelID = modelWithID.ID;
-      const modelNode = modelNodes[modelID];
-
-      this.setLookAtProperties(model, modelNode);
-
-      const parentConnections = connections.get(modelID)?.parents || [];
-
-      parentConnections.forEach(connection => {
-        const parent = modelMap.get(connection.ID);
-
-        if (parent !== undefined) {
-          parent.add(model);
-        }
-      });
-
-      if (model.parent === null) {
-        this.context.sceneGraph.add(model);
-      }
-    });
-
-    this.bindSkeleton(deformers.skeletons, geometryMap, modelMap);
-
-    this.addGlobalSceneSettings();
-
-    this.context.sceneGraph.traverse(node => {
-      if (node.userData.transformData) {
-        if (node.parent) {
-          node.userData.transformData.parentMatrix = node.parent.matrix;
-          node.userData.transformData.parentMatrixWorld = node.parent.matrixWorld;
-        }
-
-        const transform = generateTransform(node.userData.transformData);
-
-        node.applyMatrix4(transform);
-        node.updateWorldMatrix(true, true);
-      }
-    });
-
-    const animations = new AnimationParser(this.context).parse(null, this.context);
-
-    // if all the models where already combined in a single group, just return that
-    if (
-      this.context.sceneGraph.children.length === 1
-      && this.context.sceneGraph.children[0].type === 'Group'
-    ) {
-      const childGroup = this.context.sceneGraph.children[0] as Group;
-
-      childGroup.animations = animations;
-      this.context.sceneGraph = childGroup;
-    }
-
-    this.context.sceneGraph.animations = animations;
-
-    return {
-      animations: animations,
-      modelInfo: modelInfoMap,
-      scene: this.context.sceneGraph,
-      sceneInfo: this.sceneInfo,
-    };
-  }
-
-  // parse nodes in context.fbxTree.Objects.Model
-  parseWireFrameModels (
-    skeletons: Record<string, FBXSkeleton>,
-    geometryMap: Map<number, BufferGeometry>,
-    materialMap: Map<number, MeshStandardMaterial>,
-    geoInfoMap: Map<number, ModelInfo>,
-  ) {
-    const modelMap: Map<number, Object3D> = new Map();
-    const fbxTree = this.context.fbxTree;
-    const connections = this.context.connections;
-    const modelInfoMap: Record<string, ModelInfo> = {};
-
-    if (!fbxTree || !fbxTree.Objects || !connections) {
-      throw new Error('Global FBXTree or fbxTree.Objects or global connections is undefined');
-    }
-    const modelNodes = fbxTree.Objects.Model;
-
-    for (const nodeID in modelNodes) {
-      const id = parseInt(nodeID);
-      const node = modelNodes[nodeID];
-      const relationships = connections.get(id) || { children: [], parents: [] };
-      const nodeAttributeName = node.attrName || '';
-      let model: Object3D | null = this.buildSkeleton(
-        relationships,
-        skeletons,
-        id,
-        nodeAttributeName,
-      );
-
-      if (!model) {
-        switch (node.attrType) {
-          case 'Camera':
-            model = this.createCamera(relationships);
-
-            break;
-          case 'Light':
-            model = this.createLight(relationships);
-
-            break;
-          case 'Mesh':
-            {
-              const result = this.createMesh(relationships, geometryMap, materialMap, geoInfoMap);
-
-              model = result.model;
-              modelInfoMap[model.name] = result.modelInfo;
-            }
-
-            break;
-          case 'NurbsCurve':
-            model = this.createCurve(relationships, geometryMap);
-
-            break;
-          case 'LimbNode':
-          case 'Root':
-            model = new Bone();
-
-            break;
-          case 'Null':
-          default:
-            model = new Group();
-
-            break;
-        }
-
-        model.name = node.attrName ? PropertyBinding.sanitizeNodeName(node.attrName) : '';
-        model.userData.originalName = node.attrName;
-
-        (model as Object3DWithID).ID = id;
-      }
-
-      if (model === null) {
-        throw new Error('THREE.FBXLoader: Model is null');
-      }
-      this.getTransformData(model, node);
-      modelMap.set(id, model);
-    }
-
-    return { modelInfoMap, modelMap };
-  }
-
-  buildSkeleton (
-    relationships: FBXConnectionNode,
-    skeletons: Record<string, FBXSkeleton>,
-    id: number,
-    name: string,
-  ) {
-    let bone: Bone | null = null;
-    const buildSkeletons = skeletons;
-
-    relationships.parents.forEach(parent => {
-      Object.entries(buildSkeletons).forEach(([, skeleton]) => {
-        skeleton.rawBones.forEach((rawBone, i) => {
-          if (rawBone.ID === parent.ID) {
-            const subBone = bone;
-
-            bone = new Bone();
-
-            bone.matrixWorld.copy(rawBone.transformLink);
-
-            // set name and id here - otherwise in cases where "subBone" is created it will not have a name / id
-
-            bone.name = name ? PropertyBinding.sanitizeNodeName(name) : '';
-            bone.userData.originalName = name;
-            (bone as BoneWithID).ID = id;
-
-            skeleton.bones[i] = bone;
-
-            // In cases where a bone is shared between multiple meshes
-            // duplicate the bone here and add it as a child of the first bone
-            if (subBone !== null) {
-              bone.add(subBone);
-            }
-          }
-        });
-      });
-    });
-
-    return bone;
-  }
-
-  // create a PerspectiveCamera or OrthographicCamera
-  createCamera (relationships: FBXConnectionNode) {
-    const fbxTree = this.context.fbxTree;
-    let model;
-    let cameraAttribute: FBXLightNodeAttribute | undefined;
-
-    if (!fbxTree || !fbxTree.Objects) {
-      throw new Error('Global FBXTree or fbxTree.Objects is undefined');
-    }
-
-    relationships.children.forEach(child => {
-      const nodeAttribute = fbxTree.Objects?.NodeAttribute;
-
-      if (!nodeAttribute) {
-        throw new Error('NodeAttribute is undefined');
-      }
-
-      const attr = nodeAttribute[child.ID];
-
-      if (attr !== undefined) {
-        cameraAttribute = attr;
-      }
-    });
-
-    if (!cameraAttribute) {
-      model = new Object3D();
-    } else {
-      let type = 0;
-
-      if (
-        cameraAttribute.CameraProjectionType !== undefined
-        && cameraAttribute.CameraProjectionType.value === 1
-      ) {
-        type = 1;
-      }
-
-      let nearClippingPlane = 1;
-
-      if (cameraAttribute.NearPlane !== undefined) {
-        const nearPlane = cameraAttribute.NearPlane.value;
-
-        if (typeof nearPlane !== 'number') {
-          throw new Error('THREE.FBXLoader: Invalid near plane value');
-        }
-
-        nearClippingPlane = nearPlane / 1000;
-      }
-
-      let farClippingPlane = 1000;
-
-      if (cameraAttribute.FarPlane !== undefined) {
-        const farPlane = cameraAttribute.FarPlane.value;
-
-        if (typeof farPlane !== 'number') {
-          throw new Error('THREE.FBXLoader: Invalid far plane value');
-        }
-        // FBX stores the far plane in millimeters
-
-        farClippingPlane = farPlane / 1000;
-      }
-
-      let width = window.innerWidth;
-      let height = window.innerHeight;
-
-      if (cameraAttribute.AspectWidth !== undefined && cameraAttribute.AspectHeight !== undefined) {
-        const aspectWidth = cameraAttribute.AspectWidth.value;
-        const aspectHeight = cameraAttribute.AspectHeight.value;
-
-        if (typeof aspectWidth !== 'number' || typeof aspectHeight !== 'number') {
-          throw new Error('THREE.FBXLoader: Invalid aspect width or height value');
-        }
-        width = aspectWidth;
-        height = aspectHeight;
-      }
-
-      const aspect = width / height;
-
-      let fov = 45;
-
-      if (cameraAttribute.FieldOfView !== undefined) {
-        const fieldOfView = cameraAttribute.FieldOfView.value;
-
-        if (typeof fieldOfView !== 'number') {
-          throw new Error('THREE.FBXLoader: Invalid field of view value');
-        }
-        // FBX stores the field of view in radians
-        fov = fieldOfView;
-      }
-
-      const focalLength = cameraAttribute.FocalLength ? cameraAttribute.FocalLength.value : null;
-
-      if (typeof focalLength !== 'number' && focalLength !== null) {
-        throw new Error('THREE.FBXLoader: Invalid focal length value');
-      }
-      switch (type) {
-        case 0: // Perspective
-          model = new PerspectiveCamera(fov, aspect, nearClippingPlane, farClippingPlane);
-          if (focalLength !== null) {
-            model.setFocalLength(focalLength);
-          }
-
-          break;
-        case 1: // Orthographic
-          console.warn('THREE.FBXLoader: Orthographic cameras not supported yet.');
-          model = new Object3D();
-
-          break;
-        default:
-          console.warn('THREE.FBXLoader: Unknown camera type ' + type + '.');
-          model = new Object3D();
-
-          break;
-      }
-    }
-
-    return model;
-  }
-
-  // Create a DirectionalLight, PointLight or SpotLight
-  createLight (relationships: FBXConnectionNode) {
-    const fbxTree = this.context.fbxTree;
-
-    if (!fbxTree || !fbxTree.Objects) {
-      throw new Error('Global FBXTree or fbxTree.Objects is undefined');
-    }
-    let model;
-    let lightAttribute: FBXLightNodeAttribute | undefined;
-
-    const nodeAttribute = fbxTree.Objects.NodeAttribute;
-
-    if (!nodeAttribute) {
-      throw new Error('NodeAttribute is undefined');
-    }
-    relationships.children.forEach(child => {
-      const attr = nodeAttribute[child.ID];
-
-      if (attr !== undefined) {
-        lightAttribute = attr;
-      }
-    });
-
-    if (lightAttribute === undefined) {
-      model = new Object3D();
-    } else {
-      let type;
-
-      // LightType can be undefined for Point lights
-      if (lightAttribute.LightType === undefined) {
-        type = 0;
-      } else {
-        type = lightAttribute.LightType.value;
-      }
-
-      let color: Color = new Color(0xffffff);
-
-      if (lightAttribute.Color !== undefined) {
-        if (typeof lightAttribute.Color.value !== 'object') {
-          throw new Error('THREE.FBXLoader: Invalid light color value');
-        }
-
-        color = ColorManagement.toWorkingColorSpace(
-          new Color().fromArray(lightAttribute.Color.value),
-          SRGBColorSpace,
-        );
-      }
-
-      if (typeof lightAttribute.Intensity?.value !== 'number') {
-        throw new Error('THREE.FBXLoader: Invalid light intensity value');
-      }
-      let intensity
-        = lightAttribute.Intensity === undefined ? 1 : lightAttribute.Intensity.value / 100;
-
-      // light disabled
-      if (
-        lightAttribute.CastLightOnObject !== undefined
-        && lightAttribute.CastLightOnObject.value === 0
-      ) {
-        intensity = 0;
-      }
-
-      let distance = 0;
-
-      if (lightAttribute.FarAttenuationEnd !== undefined) {
-        if (
-          lightAttribute.EnableFarAttenuation !== undefined
-          && lightAttribute.EnableFarAttenuation.value === 0
-        ) {
-          distance = 0;
-        } else {
-          if (typeof lightAttribute.FarAttenuationEnd.value !== 'number') {
-            throw new Error('THREE.FBXLoader: Invalid light far attenuation end value');
-          }
-
-          distance = lightAttribute.FarAttenuationEnd.value;
-        }
-      }
-
-      // TODO: could this be calculated linearly from FarAttenuationStart to FarAttenuationEnd?
-      const decay = 1;
-
-      switch (type) {
-        case 0: // Point
-          model = new PointLight(color, intensity, distance, decay);
-
-          break;
-        case 1: // Directional
-          model = new DirectionalLight(color, intensity);
-
-          break;
-        case 2: // Spot
-          {
-            let angle = Math.PI / 3;
-
-            if (lightAttribute.InnerAngle !== undefined) {
-              const innerAngle = lightAttribute.InnerAngle.value;
-
-              if (typeof innerAngle !== 'number') {
-                throw new Error('THREE.FBXLoader: Invalid light inner angle value');
-              }
-              angle = MathUtils.degToRad(innerAngle);
-            }
-
-            let penumbra = 0;
-
-            if (lightAttribute.OuterAngle !== undefined) {
-              const outerAngle = lightAttribute.OuterAngle.value;
-
-              if (typeof outerAngle !== 'number') {
-                throw new Error('THREE.FBXLoader: Invalid light outer angle value');
-              }
-
-              // TODO: this is not correct - FBX calculates outer and inner angle in degrees
-              // with OuterAngle > InnerAngle && OuterAngle <= Math.PI
-              // while three.js uses a penumbra between (0, 1) to attenuate the inner angle
-              penumbra = MathUtils.degToRad(outerAngle);
-              penumbra = Math.max(penumbra, 1);
-            }
-
-            model = new SpotLight(color, intensity, distance, angle, penumbra, decay);
-          }
-
-          break;
-        default:
-          console.warn(
-            'THREE.FBXLoader: Unknown light type '
-            + lightAttribute.LightType?.value
-            + ', defaulting to a PointLight.',
-          );
-          model = new PointLight(color, intensity);
-
-          break;
-      }
-
-      if (lightAttribute.CastShadows !== undefined && lightAttribute.CastShadows.value === 1) {
-        model.castShadow = true;
-      }
-    }
-
-    return model;
-  }
-
-  createMesh (
-    relationships: FBXConnectionNode,
-    geometryMap: Map<number, BufferGeometry>,
-    materialMap: Map<number, MeshStandardMaterial>,
-    geoInfoMap: Map<number, ModelInfo>,
-  ) {
-    let model: Mesh | SkinnedMesh;
-    let geometry: BufferGeometry | undefined;
-    let material: MeshStandardMaterial | MeshStandardMaterial[] | null = null;
-    let modelInfo: GeometryInfo = {
-      isPBR: false,
-      isTextured: false,
-      isUVMapped: false,
-      polygons: 0,
-      quads: 0,
-      triangles: 0,
-      vertices: 0,
-    };
-    const materials: MeshStandardMaterial[] = [];
-
-    // get geometry and materials(s) from connections
-    relationships.children.forEach(child => {
-      if (geometryMap.has(child.ID)) {
-        geometry = geometryMap.get(child.ID);
-        modelInfo = geoInfoMap.get(child.ID) || modelInfo;
-      }
-
-      if (materialMap.has(child.ID)) {
-        const material = materialMap.get(child.ID);
-
-        if (material) {
-          materials.push(material);
-        }
-      }
-    });
-
-    if (materials.length > 1) {
-      material = materials;
-    } else if (materials.length > 0) {
-      material = materials[0];
-    } else {
-      const defaultMaterial = materialMap.get(this.defaultMaterialIndex);
-
-      if (defaultMaterial) {
-        material = defaultMaterial;
-        materials.push(defaultMaterial);
-      }
-    }
-
-    if (material) {
-      if (Array.isArray(material)) {
-        // 如果是材质数组，检查第一个材质
-        const firstMaterial = material[0];
-
-        if (firstMaterial) {
-          modelInfo.isPBR = Boolean(firstMaterial.metalnessMap) && Boolean(firstMaterial.roughnessMap);
-          modelInfo.isTextured = Boolean(firstMaterial.map);
-        }
-      } else {
-        // 单个材质
-        modelInfo.isPBR = Boolean(material.metalnessMap) && Boolean(material.roughnessMap);
-        modelInfo.isTextured = Boolean(material.map);
-      }
-      modelInfo.isUVMapped = Boolean(geometry?.attributes.uv);
-    }
-
-    if (!geometry) {
-      throw new Error('THREE.FBXLoader: No geometry found for mesh.');
-    }
-
-    if ('color' in geometry.attributes) {
-      materials.forEach(material => {
-        material.vertexColors = true;
-      });
-    }
-
-    // Sanitization: If geometry has groups, then it must match the provided material array.
-    // If not, we need to clean up the `group.materialIndex` properties inside the groups and point at a (new) default material.
-    // This isn't well defined; Unity creates default material, while Blender implicitly uses the previous material in the list.
-    if (geometry.groups.length > 0) {
-      let needsDefaultMaterial = false;
-
-      for (let i = 0, il = geometry.groups.length; i < il; i++) {
-        const group = geometry.groups[i] as GeometryGroup;
-        const materialIndex = group.materialIndex;
-
-        if (materialIndex == null || materialIndex < 0 || materialIndex >= materials.length) {
-          group.materialIndex = materials.length;
-          needsDefaultMaterial = true;
-        }
-      }
-
-      if (needsDefaultMaterial) {
-        const defaultMaterial = materialMap.get(this.defaultMaterialIndex);
-
-        if (defaultMaterial) {
-          materials.push(defaultMaterial);
-        }
-      }
-    }
-
-    if ((geometry as GeometryWithDeformer).FBX_Deformer) {
-      const skinnedMesh = new SkinnedMesh(geometry, material || undefined);
-
-      skinnedMesh.normalizeSkinWeights();
-      model = skinnedMesh;
-    } else {
-      model = new Mesh(geometry, material || undefined);
-    }
-
-    // 更新场景统计信息
-    this.sceneInfo.polygons += modelInfo.polygons;
-    this.sceneInfo.quads += modelInfo.quads;
-    this.sceneInfo.triangles += modelInfo.triangles;
-    this.sceneInfo.vertices += modelInfo.vertices;
-
-    if (!modelInfo.isPBR) {
-      this.sceneInfo.isPBR = false;
-    }
-    if (!modelInfo.isTextured) {
-      this.sceneInfo.isTextured = false;
-    }
-    if (!modelInfo.isUVMapped) {
-      this.sceneInfo.isUVMapped = false;
-    }
-
-    return { model, modelInfo };
-  }
-
-  createCurve (relationships: FBXConnectionNode, geometryMap: Map<number, BufferGeometry>) {
-    const geometry = relationships.children.reduce<BufferGeometry | null>((geo, child) => {
-      if (geometryMap.has(child.ID)) {
-        geo = geometryMap.get(child.ID) ?? null;
-      }
-      if (!geo) {
-        throw new Error('THREE.FBXLoader: No geometry found for curve.');
-      }
-
-      return geo;
-    }, null);
-
-    // FBX does not list materials for Nurbs lines, so we'll just put our own in here.
-    const material = new LineBasicMaterial({
-      color: 0x3300ff,
-      linewidth: 1,
-      name: Loader.DEFAULT_MATERIAL_NAME,
-    });
-
-    if (geometry === null) {
-      throw new Error('THREE.FBXLoader: No geometry found for curve.');
-    }
-
-    return new Line(geometry, material);
-  }
-
-  // parse the model node for transform data
-  getTransformData (model: Object3D, modelNode: FBXModelNode) {
-    const transformData: UserDataTransform = {};
-
-    if ('InheritType' in modelNode) {
-      const inheritTypeValue = extractMaterialValue(modelNode.InheritType);
-
-      if (typeof inheritTypeValue === 'string' || typeof inheritTypeValue === 'number') {
-        transformData.inheritType = parseInt(String(inheritTypeValue));
-      }
-    }
-
-    if ('RotationOrder' in modelNode) {
-      const value = modelNode.RotationOrder.value;
-
-      if (typeof value !== 'number') {
-        throw new Error('THREE.FBXLoader: Invalid rotation order value');
-      }
-      transformData.eulerOrder = getEulerOrder(value) as EulerOrder;
-    } else {
-      transformData.eulerOrder = getEulerOrder(0) as EulerOrder;
-    }
-
-    if ('Lcl_Translation' in modelNode) {
-      // 如果scale需要从100调整为1，position也需要相应调整
-      const translationArray = extractNumberArray(modelNode.Lcl_Translation.value);
-
-      if (translationArray && translationArray.length >= 3) {
-        let scaleAdjustmentFactor = 1;
-
-        if ('Lcl_Scaling' in modelNode) {
-          const scaleArray = extractNumberArray(modelNode.Lcl_Scaling.value);
-
-          if (scaleArray && scaleArray.length >= 3) {
-            const scaleX = scaleArray[0];
-            const scaleY = scaleArray[1];
-            const scaleZ = scaleArray[2];
-
-            if (scaleX === 100 || scaleY === 100 || scaleZ === 100) {
-              scaleAdjustmentFactor = 100;
-            }
-          }
-        }
-
-        transformData.translation = [
-          translationArray[0] / scaleAdjustmentFactor,
-          translationArray[1] / scaleAdjustmentFactor,
-          translationArray[2] / scaleAdjustmentFactor,
-        ];
-      }
-    }
-
-    if ('PreRotation' in modelNode) {
-      const valueArray = extractNumberArray(modelNode.PreRotation.value);
-
-      if (valueArray && valueArray.length >= 3) {
-        transformData.preRotation = [
-          Number(valueArray[0].toFixed(4)),
-          Number(valueArray[1].toFixed(4)),
-          Number(valueArray[2].toFixed(4)),
-        ];
-      }
-    }
-
-    if ('Lcl_Rotation' in modelNode) {
-      const valueArray = extractNumberArray(modelNode.Lcl_Rotation.value);
-
-      if (valueArray && valueArray.length >= 3) {
-        transformData.rotation = [
-          Number(valueArray[0].toFixed(4)),
-          Number(valueArray[1].toFixed(4)),
-          Number(valueArray[2].toFixed(4)),
-        ];
-      }
-    }
-
-    if ('PostRotation' in modelNode) {
-      const valueArray = extractNumberArray(modelNode.PostRotation.value);
-
-      if (valueArray && valueArray.length >= 3) {
-        transformData.postRotation = [
-          Number(valueArray[0].toFixed(4)),
-          Number(valueArray[1].toFixed(4)),
-          Number(valueArray[2].toFixed(4)),
-        ];
-      }
-    }
-
-    if ('Lcl_Scaling' in modelNode) {
-      const scaleArray = extractNumberArray(modelNode.Lcl_Scaling.value);
-
-      if (scaleArray && scaleArray.length >= 3) {
-        transformData.scale = [
-          scaleArray[0] === 100 ? 1 : scaleArray[0],
-          scaleArray[1] === 100 ? 1 : scaleArray[1],
-          scaleArray[2] === 100 ? 1 : scaleArray[2],
-        ];
-      }
-    }
-
-    if ('ScalingOffset' in modelNode) {
-      const scalingOffsetArray = extractNumberArray(modelNode.ScalingOffset.value);
-
-      if (scalingOffsetArray) {
-        transformData.scalingOffset = scalingOffsetArray as [number, number, number];
-      }
-    }
-
-    if ('ScalingPivot' in modelNode) {
-      const scalingPivotArray = extractNumberArray(modelNode.ScalingPivot.value);
-
-      if (scalingPivotArray) {
-        transformData.scalingPivot = scalingPivotArray as [number, number, number];
-      }
-    }
-
-    if ('RotationOffset' in modelNode) {
-      const rotationOffsetArray = extractNumberArray(modelNode.RotationOffset.value);
-
-      if (rotationOffsetArray) {
-        transformData.rotationOffset = rotationOffsetArray as [number, number, number];
-      }
-    }
-
-    if ('RotationPivot' in modelNode) {
-      const rotationPivotArray = extractNumberArray(modelNode.RotationPivot.value);
-
-      if (rotationPivotArray) {
-        transformData.rotationPivot = rotationPivotArray as [number, number, number];
-      }
-    }
-
-    model.userData.transformData = transformData;
-  }
-
-  setLookAtProperties (model: Object3D, modelNode: FBXModelNode) {
-    const connections = this.context.connections;
-    const fbxTree = this.context.fbxTree;
-
-    if (!connections || !fbxTree) {
-      throw new Error('Global connections or FBX tree is undefined');
-    }
-    if ('LookAtProperty' in modelNode) {
-      const modelWithID = model as Object3DWithID;
-      const children = connections.get(modelWithID.ID)?.children ?? [];
-
-      children.forEach(child => {
-        if (child.relationship === 'LookAtProperty') {
-          const modelNode = fbxTree.Objects?.Model;
-
-          if (!modelNode) {
-            throw new Error('Model node is undefined');
-          }
-          const lookAtTarget = modelNode[child.ID];
-
-          if ('Lcl_Translation' in lookAtTarget) {
-            const posArray = extractNumberArray(lookAtTarget.Lcl_Translation.value);
-
-            if (posArray && posArray.length >= 3) {
-              let scaleAdjustmentFactor = 1;
-
-              if ('Lcl_Scaling' in lookAtTarget) {
-                const scaleArray = extractNumberArray(lookAtTarget.Lcl_Scaling.value);
-
-                if (scaleArray && scaleArray.length >= 3) {
-                  const scaleX = scaleArray[0];
-                  const scaleY = scaleArray[1];
-                  const scaleZ = scaleArray[2];
-
-                  if (scaleX === 100 || scaleY === 100 || scaleZ === 100) {
-                    scaleAdjustmentFactor = 100;
-                  }
-                }
-              }
-
-              // 调整position以保持一致性
-              const pos = [
-                posArray[0] / scaleAdjustmentFactor,
-                posArray[1] / scaleAdjustmentFactor,
-                posArray[2] / scaleAdjustmentFactor,
-              ];
-
-              // DirectionalLight, SpotLight
-              if (hasTarget(model)) {
-                model.target.position.fromArray(pos);
-                const sceneGraph = this.context.sceneGraph;
-
-                sceneGraph.add(model.target);
-              } else {
-                // Cameras and other Object3Ds
-                model.lookAt(new Vector3().fromArray(pos));
-              }
-            }
-          }
-        }
-      });
-    }
-  }
-
-  bindSkeleton (
-    skeletons: Record<string, FBXSkeleton>,
-    geometryMap: Map<number, BufferGeometry>,
-    modelMap: Map<number, Object3D>,
-  ) {
-    const fbxTree = this.context.fbxTree;
-    const connections = this.context.connections;
-
-    if (!fbxTree || !connections) {
-      throw new Error('Global fbxTree or connections is undefined');
-    }
-
-    const bindMatrices = this.parsePoseNodes();
-
-    for (const ID in skeletons) {
-      const skeleton = skeletons[ID];
-
-      const parents = connections.get(parseInt(skeleton.ID))?.parents || [];
-
-      parents.forEach(parent => {
-        if (geometryMap.has(parent.ID)) {
-          const geoID = parent.ID;
-          const geoRelationships = connections.get(geoID) || { parents: [] };
-
-          geoRelationships.parents.forEach(geoConnParent => {
-            if (modelMap.has(geoConnParent.ID)) {
-              const model = modelMap.get(geoConnParent.ID);
-
-              if (model && model.type === 'SkinnedMesh') {
-                const skinnedMesh = model as SkinnedMesh;
-
-                skinnedMesh.bind(new Skeleton(skeleton.bones), bindMatrices[geoConnParent.ID]);
-              }
-            }
-          });
-        }
-      });
-    }
-  }
-
-  parsePoseNodes () {
-    const bindMatrices: Record<string, Matrix4> = {};
-    const fbxTree = this.context.fbxTree;
-    const connections = this.context.connections;
-
-    if (!fbxTree || !connections || !fbxTree.Objects) {
-      throw new Error('Global fbxTree or connections is undefined');
-    }
-
-    if ('Pose' in fbxTree.Objects) {
-      const BindPoseNode = fbxTree.Objects.Pose;
-
-      for (const nodeID in BindPoseNode) {
-        const bindPoseNode = BindPoseNode[nodeID];
-
-        if (bindPoseNode.attrType === 'BindPose' && bindPoseNode.NbPoseNodes > 0) {
-          const poseNodes = bindPoseNode.PoseNode;
-
-          if (Array.isArray(poseNodes)) {
-            poseNodes.forEach(poseNode => {
-              const nodeId = extractNodeId(poseNode);
-              const matrixArray = extractMatrixArray(poseNode);
-
-              bindMatrices[nodeId || 0] = new Matrix4().fromArray(matrixArray);
-            });
-          } else {
-            const node = extractNodeId(poseNodes);
-            const matrixArray = extractMatrixArray(poseNodes);
-
-            if (!node) {
-              throw new Error('THREE.FBXLoader: No node found for poseNode.');
-            }
-
-            bindMatrices[node] = new Matrix4().fromArray(matrixArray);
-          }
-        }
-      }
-    }
-
-    return bindMatrices;
-  }
-
-  addGlobalSceneSettings () {
-    const fbxTree = this.context.fbxTree;
-
-    if (!fbxTree || !fbxTree.GlobalSettings) {
-      throw new Error('Global FBXTree is undefined');
-    }
-    if ('GlobalSettings' in fbxTree) {
-      if ('AmbientColor' in fbxTree.GlobalSettings) {
-        // Parse ambient color - if it's not set to black (default), create an ambient light
-
-        const ambientColorArray = extractNumberArray(fbxTree.GlobalSettings.AmbientColor.value);
-
-        if (ambientColorArray && ambientColorArray.length >= 3) {
-          const r = ambientColorArray[0];
-          const g = ambientColorArray[1];
-          const b = ambientColorArray[2];
-
-          if (r !== 0 || g !== 0 || b !== 0) {
-            const color = new Color().setRGB(r, g, b);
-
-            this.context.sceneGraph.add(new AmbientLight(color, 1));
-          }
-        }
-      }
-
-      if ('UnitScaleFactor' in fbxTree.GlobalSettings) {
-        this.context.sceneGraph.userData.unitScaleFactor = fbxTree.GlobalSettings.UnitScaleFactor.value;
-      }
-    }
-  }
-
-}
